tidy files and replace with correct version
