+++ VirtualTourist/VirtualTourist/VirtualTourist/VirtualTourist.xcodeproj/project.xcworkspace/xcshareddata/VirtualTourist.xccheckout
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
+	<false/>
+	<key>IDESourceControlProjectIdentifier</key>
+	<string>A1255DA0-7274-4E7A-8534-80B8DA085796</string>
+	<key>IDESourceControlProjectName</key>
+	<string>VirtualTourist</string>
+	<key>IDESourceControlProjectOriginsDictionary</key>
+	<dict>
+		<key>240DBD38FD10C87990CBC4932A98F835B686EA3F</key>
+		<string>https://bitbucket.org/grantkemp/udacity-training-submission.git</string>
+		<key>3CD6DB2D91D7599DF412C0928ADBF1FD9753E657</key>
+		<string>https://bitbucket.org/grantkemp/udacity-training-submission.git</string>
+	</dict>
+	<key>IDESourceControlProjectPath</key>
+	<string>VirtualTourist/VirtualTourist/VirtualTourist.xcodeproj</string>
+	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
+	<dict>
+		<key>240DBD38FD10C87990CBC4932A98F835B686EA3F</key>
+		<string>../../../..</string>
+		<key>3CD6DB2D91D7599DF412C0928ADBF1FD9753E657</key>
+		<string>../../../../..</string>
+	</dict>
+	<key>IDESourceControlProjectURL</key>
+	<string>https://bitbucket.org/grantkemp/udacity-training-submission.git</string>
+	<key>IDESourceControlProjectVersion</key>
+	<integer>111</integer>
+	<key>IDESourceControlProjectWCCIdentifier</key>
+	<string>240DBD38FD10C87990CBC4932A98F835B686EA3F</string>
+	<key>IDESourceControlProjectWCConfigurations</key>
+	<array>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>3CD6DB2D91D7599DF412C0928ADBF1FD9753E657</string>
+			<key>IDESourceControlWCCName</key>
+			<string></string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>240DBD38FD10C87990CBC4932A98F835B686EA3F</string>
+			<key>IDESourceControlWCCName</key>
+			<string>Submission</string>
+		</dict>
+	</array>
+</dict>
+</plist>
