+++ VirtualTourist/VirtualTourist/VirtualTourist/VirtualTourist/FlickrConfig.swift
