+++ VirtualTourist/VirtualTourist/VirtualTourist/VirtualTourist/FlickrConvenience.swift
