+++ VirtualTourist/VirtualTourist/VirtualTourist/VirtualTourist/Images.xcassets/AppIcon.appiconset/Contents.json
