+++ VirtualTourist/VirtualTourist/VirtualTourist/VirtualTourist/Images.xcassets/PlaceholderImage.imageset/Contents.json
