+++ VirtualTourist/VirtualTourist/VirtualTourist/VirtualTourist/MyPlayground.playground/Sources/SupportCode.swift
