+++ VirtualTourist/VirtualTourist/VirtualTourist/VirtualTourist/PhotoHelper.swift
