+++ VirtualTourist/VirtualTourist/VirtualTourist/VirtualTourist/PinViewCollectionViewCell.swift
